结束了哈哈哈

+++ GPText/GPText/NSString+Emoji.m
@@ -1,10 +1,4 @@
-//
-//  NSString+Emoji.m
-//  黑马微博
-//
-//  Created by MJ Lee on 14/7/12.
-//  Copyright (c) 2014年 heima. All rights reserved.
-//
+
 
 #import "NSString+Emoji.h"
 #define EmojiCodeToSymbol(c) ((((0x808080F0 | (c & 0x3F000) >> 4) | (c & 0xFC0) << 10) | (c & 0x1C0000) << 18) | (c & 0x3F) << 24)
@@ -23,7 +17,7 @@
 + (NSString *)emojiWithStringCode:(NSString *)stringCode
 {
     char *charCode = (char *)stringCode.UTF8String;
-    int intCode = strtol(charCode, NULL, 16);
+    NSInteger intCode = strtol(charCode, NULL, 16);
     return [self emojiWithIntCode:intCode];
 }
